Commit May 7 2014 -- List Integer
This commit is adding the progress I've made on a list-based number.
This will allow us to avoid overflows by breaking the number down by digits.
Each digit will be a node in a linked list that makes up the number.
This way the length of the number is only limited by the amount of RAM available
to the system.

The spike folder also has a simple makefile for testing the integer.

Simple changes and additions have been made to the mathx and rsa files.
rsa almost has a working keygen. Once the integer library is finished
the keygen will work fine.