Commit May 8 2014 -- Creating Integers From Character strings
I have a "constructor" for an integer working.
It takes a character string and converts it into an integer.
This way, we can have infinitly long integers since strings are not
restricted in size. This way, we can excercise more of the potential.

I am going to try implementing the keygen without the new integer
and see if we can avoid overflows. If not, I'll come back to this.